refactor(packages): consolidate markdown plugins and rename shared to data
- Merge @zlliang/rehype and @zlliang/remark into a new @zlliang/markdown
  package with subpath exports ./remark and ./rehype
- Rename @zlliang/shared to @zlliang/data
- Update theme integration, websites, and skill references accordingly
