+++ .agents/skills/website-development/references/commands.md
@@ -23,8 +23,8 @@ Run commands from the repository root.
 
 ## Shared imports
 
-- Import shared remark plugins from `@zlliang/remark`.
-- Import shared rehype plugins from `@zlliang/rehype`.
+- Import shared remark plugins from `@zlliang/markdown/remark`.
+- Import shared rehype plugins from `@zlliang/markdown/rehype`.
 - Import shared styles, utilities, and components from `@zlliang/theme/*`.
 
 ## Validation
