+++ packages/data/assets/logo-mesh.png
