+++ packages/data/assets/logo-muse.png
