+++ packages/data/package.json
@@ -1,5 +1,5 @@
 {
-  "name": "@zlliang/shared",
+  "name": "@zlliang/data",
   "version": "0.0.0",
   "private": true,
   "type": "module",
