--- packages/remark/package.json
@@ -1,16 +1,0 @@
-{
-  "name": "@zlliang/remark",
-  "version": "0.0.0",
-  "private": true,
-  "type": "module",
-  "exports": {
-    ".": "./src/index.ts"
-  },
-  "dependencies": {
-    "remark-cjk-friendly": "^2.0.1",
-    "remark-flexible-code-titles": "^1.3.3"
-  },
-  "devDependencies": {
-    "typescript": "^6.0.3"
-  }
-}
