+++ packages/theme/src/components/SiteCard.astro
@@ -1,10 +1,10 @@
 ---
 import { Image } from "astro:assets"
 
-import { siteData } from "@zlliang/shared"
+import { siteData } from "@zlliang/data"
 import { cn } from "../styles"
 
-import type { Site, Locale } from "@zlliang/shared"
+import type { Site, Locale } from "@zlliang/data"
 
 interface Props {
   site: Site
