+++ websites/home/package.json
@@ -10,7 +10,7 @@
   "dependencies": {
     "@vercel/analytics": "^2.0.1",
     "@vercel/speed-insights": "^2.0.0",
-    "@zlliang/shared": "workspace:*",
+    "@zlliang/data": "workspace:*",
     "@zlliang/theme": "workspace:*"
   },
   "devDependencies": {
