+++ websites/mesh/package.json
@@ -10,7 +10,7 @@
   "dependencies": {
     "@vercel/analytics": "^2.0.1",
     "@vercel/speed-insights": "^2.0.0",
-    "@zlliang/shared": "workspace:*",
+    "@zlliang/data": "workspace:*",
     "@zlliang/theme": "workspace:*",
     "date-fns": "^4.1.0",
     "lodash-es": "^4.18.1",
@@ -23,8 +23,6 @@
     "@iconify/tailwind4": "^1.2.3",
     "@tailwindcss/vite": "^4.2.4",
     "@types/lodash-es": "^4.17.12",
-    "@zlliang/rehype": "workspace:*",
-    "@zlliang/remark": "workspace:*",
     "astro": "^6.1.9",
     "tailwindcss": "^4.2.4",
     "typescript": "^6.0.3"
